refactor: move remaining batch methods into dataloader service
